+++ Expenses.BL/Entities/Operation.cs
@@ -4,17 +4,17 @@ namespace Expenses.BL.Entities
 {
     public class Operation : IUnique
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public DateTime OperationTime { get; set; }
 
-        public int UserId { get; set; }
+        public long UserId { get; set; }
 
-        public int AccountId { get; set; }
+        public long AccountId { get; set; }
 
         public double Amount { get; set; }
 
-        public int ExpenseItemId { get; set; }
+        public long ExpenseItemId { get; set; }
 
         public string Comment { get; set; }
     }
